+++ objc2/src/rc/test_object.rs
@@ -5,7 +5,7 @@ use core::ptr;
 use super::{Id, Owned};
 use crate::foundation::{NSObject, NSZone};
 use crate::runtime::Bool;
-use crate::{declare_class, msg_send, msg_send_bool, ClassType};
+use crate::{declare_class, msg_send, ClassType};
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
 pub(crate) struct ThreadTestData {
@@ -122,12 +122,12 @@ declare_class!(
         #[sel(_tryRetain)]
         unsafe fn try_retain(&self) -> Bool {
             TEST_DATA.with(|data| data.borrow_mut().try_retain += 1);
-            let res = unsafe { msg_send_bool![super(self), _tryRetain] };
+            let res: bool = unsafe { msg_send![super(self), _tryRetain] };
             if !res {
                 TEST_DATA.with(|data| data.borrow_mut().try_retain -= 1);
                 TEST_DATA.with(|data| data.borrow_mut().try_retain_fail += 1);
             }
-            Bool::from(res)
+            Bool::new(res)
         }
 
         #[sel(copyWithZone:)]